pyproject.toml
[project]
name = "shopify-partner-agent"
version = "0.4.0"
description = "MCP server that gives AI tools Partner API access: MRR, churn, retention cohorts, merchant health, and 20+ computed analytics for Shopify app developers."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "Rushikesh More" }]
keywords = [
    "mcp",
    "shopify",
    "shopify-partner",
    "partner-api",
    "mrr",
    "churn",
    "analytics",
    "revenue",
    "llm",
    "claude",
    "anthropic",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial",
    "Topic :: Software Development :: Libraries",
    "Topic :: Internet :: WWW/HTTP",
]

dependencies = [
    "mcp>=1.26.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
shopify-partner-agent = "shopify_partner_agent.server:main"

[project.urls]
Homepage = "https://shopify-partner-agent.vercel.app"
Repository = "https://github.com/rushikeshmore/shopify-partner-agent"
Issues = "https://github.com/rushikeshmore/shopify-partner-agent/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/shopify_partner_agent"]

[tool.hatch.build.targets.sdist]
include = [
    "src/shopify_partner_agent",
    "README.md",
    "LICENSE",
    "pyproject.toml",
]
exclude = [
    "**/.env",
    "**/.env.*",
    "**/__pycache__",
    "**/*.pyc",
    "tests/",
    "site/",
    "mcp-directory-submission.md",
    ".github/",
    "CLAUDE.md",
    "phase-tracker.md",
    "build-sop/",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "D",
    "UP",
    "B",
    "SIM",
    "RUF",
]
ignore = [
    "D100",
    "D104",
    "D203",
    "D212",
    "D213",
    "D417",
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["D"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
]
